+++ lib/talkable/generators/templates/app/views/example_newsletter_signup/register.html.slim
@@ -1,3 +1,4 @@
+h3 Example newsletter signup (this form will just submit an event to talkable)
 = form_tag :example_newsletter_signup do
   = email_field_tag :email
-  = submit_tag
+  = submit_tag 'SignUp'
